+++ src/apis/queries.ts
@@ -18,7 +18,7 @@ import {
   getPosePick,
   getPoseTalk,
 } from '.';
-import { FilterState } from '@/hooks/useFilterState';
+import { FilterState } from '';
 
 export const usePoseDetailQuery = (poseId: number) =>
   useQuery(['poseId', poseId], () => getPoseDetail(poseId));
