+++ src/app/(Main)/feed/FeedContent.tsx
@@ -5,7 +5,7 @@ import { PrimaryButton } from '@/components/Button';
 import EmptyCase from '@/components/Feed/EmptyCase';
 import FeedSection from '@/components/Feed/FeedSection';
 import { URL } from '@/constants';
-import useFilterState from '@/hooks/useFilterState';
+import { useFilterState } from '@/hooks';
 
 export default function FeedContent() {
   const { filterState } = useFilterState();
