+++ src/components/Modal/BottomSheet.tsx
@@ -1,7 +1,7 @@
 import { IconButton } from '../Button';
 import { AnimatedPortal } from '@/components/Portal';
 import { ICON } from '@/constants';
-import useBottomSheet from '@/hooks/useBottomSheet';
+import { useBottomSheet } from '@/hooks';
 import { StrictPropsWithChildren } from '@/types';
 
 export default function BottomSheet({ children }: StrictPropsWithChildren) {
@@ -17,7 +17,7 @@ export default function BottomSheet({ children }: StrictPropsWithChildren) {
       <AnimatedPortal
         motionProps={{ initial: { opacity: 0 }, animate: { opacity: 1 }, exit: { opacity: 0 } }}
       >
-        <div className="fixed inset-x-0 bottom-0 z-modal mx-auto max-w-layout rounded-t-16 bg-white">
+        <div className="fixed inset-x-0 bottom-0 mx-auto bg-white z-modal max-w-layout rounded-t-16">
           <div className="flex justify-end px-8 pt-12">
             <IconButton icon={ICON.close.black} onClick={closeBottomSheet} />
           </div>
